Fix css, add favicon
