+++ index.html
@@ -1,10 +1,14 @@
-<!DOCTYPE html>
+<!doctype html>
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/src/favicon.svg" />
+    <link rel="icon" href="/src/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Happinessiness App</title>
+    <meta
+      name="description"
+      content="Web app for logging your daily happiness"
+    />
+    <title>Happiness App</title>
   </head>
   <body class="h-full w-full">
     <div class="w-full h-full" id="root"></div>
