+++ src/index.css
@@ -10,31 +10,24 @@
 
 @layer components {
     h1 {
-        @apply text-4xl
-        @apply font-medium
+        @apply text-4xl font-medium;
     }
     h2 {
-        @apply text-3xl
-        @apply font-medium
+        @apply text-3xl font-medium;
     }
     h3 {
-        @apply text-2xl
-        @apply font-medium
+        @apply text-2xl font-medium;
     }
     h4 {
-        @apply text-xl
-        @apply font-medium
+        @apply text-xl font-medium;
     }
     h5 {
-        @apply text-lg
-        @apply font-medium
+        @apply text-lg font-medium;
     }
     p {
-        @apply text-base
-        @apply font-medium
+        @apply text-base font-medium;
     }
     label {
-        @apply text-sm
-        @apply font-medium
+        @apply text-sm font-medium;
     }
 }
