+++ src/pages/Entries/Entries.tsx
@@ -1,4 +1,4 @@
-import { useUser } from "../contexts/UserProvider";
+import { useUser } from "../../contexts/UserProvider";
 
 export default function Entries() {
   const { user } = useUser();
